remove apps.api.data and create new indexing task
