+++ designsafe/apps/api/tasks.py
@@ -19,120 +19,6 @@ from elasticsearch_dsl.query import Q
 
 logger = logging.getLogger(__name__)
 
-@shared_task(bind=True, max_retries=None)
-def reindex_agave(self, username, file_id, full_indexing=True,
-                  levels=1, pems_indexing=True, index_full_path=True):
-    user = get_user_model().objects.get(username=username)
-    #levels=1
-    
-    from designsafe.apps.api.data import AgaveFileManager
-    from designsafe.apps.data.managers.indexer import AgaveIndexer as AgaveFileIndexer
-    agave_fm = AgaveFileManager(user)
-    
-    if settings.DEBUG and username == 'ds_admin':
-        service_client = get_service_account_client()
-        agave_fm.agave_client = service_client
-        agave_fm.indexer = AgaveFileIndexer(agave_client=service_client)
-
-    system_id, file_user, file_path = agave_fm.parse_file_id(file_id)
-    if system_id != settings.AGAVE_STORAGE_SYSTEM:
-        file_id_comps = file_id.strip('/').split('/')
-        system_id = file_id_comps[0]
-        file_user = username
-        if len(file_id_comps) > 1:
-            file_path = os.path.join(*file_id_comps[1:])
-        else:
-            file_path = '/'
-
-    agave_fm.indexer.index(system_id, file_path, file_user,
-                           full_indexing = full_indexing,
-                           pems_indexing = pems_indexing,
-                           index_full_path = index_full_path,
-                           levels = levels)
-    #parent_path_comps = file_path.strip('/').split('/')
-    #if len(parent_path_comps) > 0:
-    #    parent_path = os.path.join(*file_path.strip('/').split('/')[:-1])
-    #    agave_fm.indexer.index(system_id, parent_path, file_user,
-    #                           levels = 1)
-
-
-@shared_task(bind=True)
-def share_agave(self, username, file_id, permissions, recursive):
-    try:
-        # n = Notification(event_type = 'data',
-        #                  status = 'INFO',
-        #                  operation = 'share_initializing',
-        #                  message = 'File sharing is initializing. Please wait...',
-        #                  user = username,
-        #                  extra = {'target_path': reverse('designsafe_data:data_depot',
-        #                                                  args=['agave', file_id])})
-        # n.save()
-        user = get_user_model().objects.get(username=username)
-
-        from designsafe.apps.api.data import AgaveFileManager
-        from designsafe.apps.api.data.agave.file import AgaveFile
-        from designsafe.apps.api.data.agave.elasticsearch.documents import Object
-        agave_fm = AgaveFileManager(user)
-        system_id, file_user, file_path = agave_fm.parse_file_id(file_id)
-
-        f = AgaveFile.from_file_path(system_id, username, file_path,
-                                     agave_client=agave_fm.agave_client)
-        f_dict = f.to_dict()
-
-        n = Notification(event_type = 'data',
-                         status = 'INFO',
-                         operation = 'share_initializing',
-                         message = 'File sharing is initializing. Please wait...',
-                         user = username,
-                         extra = f_dict)
-        n.save()
-
-        f.share(permissions, recursive)
-        #reindex_agave.apply_async(args=(self.username, file_id))
-        # self.indexer.index(system, file_path, file_user, pems_indexing=True)
-
-        esf = Object.from_file_path(system_id, username, file_path)
-        esf.share(username, permissions, recursive)
-
-        # Notify owner share completed
-        n = Notification(event_type = 'data',
-                         status = 'SUCCESS',
-                         operation = 'share_finished',
-                         message = 'File permissions were updated successfully.',
-                         user = username,
-                         extra = f_dict)
-        n.save()
-
-        # Notify users they have new shared files
-        for pem in permissions:
-            if pem['permission'] != 'NONE':
-                message = '%s shared some files with you.' % user.get_full_name()
-                n = Notification(event_type = 'data',
-                                 status = 'SUCCESS',
-                                 operation = 'share_finished',
-                                 message = message,
-                                 user = pem['user_to_share'],
-                                 extra = f_dict)
-                n.save()
-
-    except:
-        logger.error('Error sharing file/folder', exc_info=True,
-                     extra = {
-                         'username': username,
-                         'file_id': file_id,
-                         'permissions': permissions
-                     })
-        n = Notification(event_type='data',
-                         status=Notification.ERROR,
-                         operation='share_error',
-                         message='We were unable to share the specified folder/file(s). '
-                                 'Please try again...',
-                         user=username,
-                         extra={'system': system_id,
-                                'path': file_path
-                         })
-        n.save()
-
 @shared_task(bind=True)
 def box_download(self, username, src_resource, src_file_id, dest_resource, dest_file_id):
     """
