+++ designsafe/apps/auth/tasks.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.mail import send_mail
 from agavepy.agave import Agave, AgaveException
-from designsafe.apps.api.data.agave.filemanager import FileManager
+from designsafe.apps.data.tasks import reindex_agave
 from celery import shared_task
 
 from requests import HTTPError
@@ -60,12 +60,18 @@ def check_or_create_agave_home_dir(username):
                 # logger.debug('setfacl response: {}'.format(response))
 
                 # add dir to index
-                fm = FileManager(user)
-                fm.indexer.index(
-                    settings.AGAVE_STORAGE_SYSTEM,
-                    username, username,
-                    levels=1
-                )
+
+                reindex_agave.apply_async(kwargs={
+                                      'username': username,
+                                      'file_id': '{}/'.format(settings.AGAVE_STORAGE_SYSTEM)
+                                      },
+                                      queue='indexing')
+                #fm = FileManager(user)
+                #fm.indexer.index(
+                #    settings.AGAVE_STORAGE_SYSTEM,
+                #    username, username,
+                #    levels=1
+                #)
 
 
     except (HTTPError, AgaveException):
