+++ designsafe/apps/data/managers/elasticsearch.py
@@ -117,23 +117,7 @@ class FileManager(object):
 
         """
         if debug_mode == True:
-            # In local dev, corral isn't mounted so we have to download the file to get its mimetype.
-            import requests
-            client = get_service_account_client()
-            system = file_object['system']
-            path = file_object['path']
-
-            try:
-                f = client.files.download(systemId=system, filePath=path)
-                mimeType = magic.from_buffer(f.content, mime=True)
-            except requests.HTTPError as e:
-                if e.response.status_code == 501:
-                    # This is a 'not implemented' error that should only get thrown for directories.
-                    mimeType =  'text/directory'
-                elif e.response.status_code == 404:
-                    # The file cannot be retrieved.
-                    raise requests.HTTPError
-            return mimeType
+            return file_object.mimeType
 
         else:
             # In dev/prod, Corral is mounted and we can use the absolute path to get the mimetype.
